Extract out common code
